+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file. The format 
 
 ### Added
 - Added the `Movement` library, which will replace Traveler as the default method of moving creeps around in Overmind
+- Room planner updates:
+    - Room planner will now destroy or dismantle incorrectly-placed structures allowing you to change your room plan after it is built
+    - Finished support for bunkers
+
+### Changed
+- Workers now include dropped energy in list of objects they can recharge from and pick their recharge target more intelligently, accounting for other targeting workers
 
 ### Fixed
 - Fixed a bug where mining sites could get clogged if invaders died on the container outputs and dropped minerals which would not get withdrawn